fix : 로그인 객체 값을 null로 받아오던 문제 해결
