+++ src/main/java/com/example/Waffle/dto/userDto.java
@@ -1,31 +1,30 @@
-package dto;
+package com.example.Waffle.dto;
 
-import entity.userEntity;
+import com.example.Waffle.entity.userEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
 public class userDto {
-    private Long id;
+    private int id;
     private String email;
     private String password;
     private String name;
 
-    public userDto(userEntity userEntity){
-        this.id = userEntity.getId();
-        this.name = userEntity.getName();
-        this.email = userEntity.getEmail();
-        this.password = userEntity.getPassword();
+    public userDto(String email, String password, String name){
+        this.email = email;
+        this.password = password;
+        this.name = name;
     }
 
     public userEntity toEntity(){
-        return userEntity.builder()
-                .id(id)
+        userEntity userEntity = new userEntity().builder()
                 .email(email)
                 .password(password)
                 .name(name)
                 .build();
+        return userEntity;
     }
 
 }
