+++ src/main/java/com/example/Waffle/entity/userEntity.java
@@ -1,11 +1,10 @@
-package entity;
+package com.example.Waffle.entity;
 
-import dto.userDto;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.Setter;
 
 @Getter
 @Entity
@@ -32,14 +31,10 @@ public class userEntity {
     private String imgName;
 
     @Builder
-    public userEntity(Long id, String email, String password, String name, String introduction, String imgPath, String imgName){
-        this.id = id;
+    public userEntity(String email, String password, String name){
         this.email = email;
         this.password = password;
         this.name = name;
-        this.introduction = introduction;
-        this.imgPath = imgPath;
-        this.imgName = imgName;
     }
