+++ src/main/java/com/example/Waffle/repository/userRepository.java
@@ -1,10 +1,12 @@
-package repository;
+package com.example.Waffle.repository;
 
-import entity.userEntity;
+import com.example.Waffle.entity.userEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+@Repository
 public interface userRepository extends JpaRepository<userEntity, Long> {
     Optional<userEntity> findByemail(String email);
 }
