+++ src/main/resources/static/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>index</title>
+</head>
+<body>
+<h2>Hello Spring Boot!!</h2>
+<a href="/register">회원가입</a>
+</body>
+</html>
