+++ src/main/resources/static/register.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+  <title>register</title>
+</head>
+<body>
+<form action="/register" method="post">
+  이메일 : <input type="text" name="email"> <br>
+  비밀번호 : <input type="text" name="password"> <br>
+  이름 : <input type="text" name = "name"> <br>
+  <input type="submit" value="회원가입">
+</form>
+</body>
+</html>
